corrige diretórios
