--- assignments/02-area/submissions/rafael-lopes/CalculadoraPoligono.java
@@ -1,3 +1,0 @@
-public class CalculadoraPoligono {
-    public static void main (St)
-}
