+++ assignments/09-collatz/submissions/rafael-lopes/src/CollatzSequence.java
