+++ challenges/02-bank/submissions/rafael-lopes/Account.java
@@ -1,6 +1,3 @@
-package poo.challenges.bank02.submissions.rafael_lopes;
-
-
 
 import java.text.DecimalFormat;
 
@@ -9,33 +6,33 @@ public class Account {
     private int id;
     private double balance;
 
-    public Account(){
+    public Account() {
         this.id = ++accountCounter;
         this.balance = 0.0;
     }
 
-    public int getId(){
+    public int getId() {
         return id;
     }
 
-    public String getBalance(){
+    public String getBalance() {
         DecimalFormat df = new DecimalFormat("R$ #,##0.00");
         return df.format(balance);
     }
 
-    public boolean deposit (double amount){
-        if (amount > 0){
+    public boolean deposit(double amount) {
+        if (amount > 0) {
             balance += amount;
             return true;
         }
         return false;
     }
 
-    public boolean withdraw(double amount){
-        if (amount > 0 && amount <= balance){
+    public boolean withdraw(double amount) {
+        if (amount > 0 && amount <= balance) {
             balance -= amount;
             return true;
         }
         return false;
-    } 
+    }
 }
