--- challenges/bank02/submissions/pedro-dutra/src/Account.java
@@ -1,3 +1,0 @@
-public class Account {
-    
-}
