--- challenges/bank02/submissions/pedro-pignata/src/openAccount.java
@@ -1,4 +1,0 @@
-
-public class openAccount {
-
-}
